fix repo running under Jenkins
Jenkins doesn't create all the ref heads because it checks out a
specific revision but since we're using the existing checkout for repo
we need one created because it will always checkout a branch. If we
don't specify a branch to checkout it will attempt to checkout master so
we always need to specify a branch and it needs to be defined correctly
via a ref head.